Compile Sass to CSS

Compile Sass to CSS using libSass.

compile_sass(file = NULL, options = sass_options(), output = NULL,
  text = NULL)

Arguments

file
Path to .scss or .sass Sass file. Note that the libSass compiler expects .sass files to use the Sass Indented Syntax.

options
Compiler options for Sass. Please specify options using sass_options.

output
Specifies path to output file for compiled CSS.

text
Character vector of Sass code.

Value

If output = NULL, the function returns a one element character vector of the compiled CSS. If the output path is specified, the compiled CSS is written to that file and the function has no return value.

Note

file and text cannot be compiled at the same time. Please pass one or the other.

Examples

compile_sass(text = "foo { margin: 122px * .3; }")

#> foo {
#>   margin: 36.6px;
#> }