Compiler Options for Sass

Set compiler options for Sass. Used with compile_sass.

opts(precision = 5, output_style = "expanded",
  indented_syntax = FALSE, include_path = "",
  source_comments = FALSE, indent_type = "space", indent_width = 2,
  linefeed = "lf", output_path = "", source_map_file = "",
  source_map_root = "", source_map_embed = FALSE,
  source_map_contents = FALSE, omit_source_map_url = FALSE)

Arguments

precision
Number of decimal places.

output_style
Bracketing and formatting style of the CSS output. Possible styles: "nested", "expanded", "compact", and "compressed".

indented_syntax
Enables the compiler to parse Sass Indented Syntax in strings. Note that the compiler automatically overrides this option to TRUE or FALSE for files with .sass and .scss file extensions respectively.

include_path
Vector of paths used to resolve @import. Multiple paths are possible using a character vector of paths.

source_comments
Annotates CSS output with line and file comments from Sass file for debugging.

indent_type
Specifies the indent type as 'space' or 'tab'.

indent_width
Number of tabs or spaces used for indentation. Maximum 10.

linefeed
Specifies how new lines should be delimited. Possible values: 'lf', 'cr', 'lfcr', and 'crlf'.

output_path
Specifies the location of the output file. Note: this option will not write the file on disk. It is only for internal reference with the source map.

source_map_file
Specifies the location for Sass to write the source map.

source_map_root
Value will be included as source root in the source map information.

source_map_embed
Embeds the source map as a data URI.

source_map_contents
Includes the contents in the source map information.

omit_source_map_url
Disable the inclusion of source map information in the output file. Note: must specify output_path when TRUE.

Value

List of Sass compiler options to be used with compile_sass.

Examples

compile_sass(
  text = "foo { margin: 122px * .3; }",
  options = opts(output_style = "compact")
)

#> foo { margin: 36.6px; }